+++ routers/portal.js
@@ -101,7 +101,7 @@ function* platform(next) {
     const product = yield findActiveWorkflowExample(__);
     const {members} = product;
 
-    const index = members.findIndex(member => member.maill === user.mail);
+    const index = members.findIndex(member => member.mail === user.mail);
 
     Object.assign(_, {
       products: __,
@@ -123,11 +123,11 @@ function* product(next) {
     return yield next;
   }
   try {
-    const [equipmentTags, platformTags] = yield [findTags("设备"), findTags("产品")];
+    const [equipmentTags, platformTags, user] = yield [findTags("设备"), findTags("产品"), this.authorized];
     const _ = {
       title: "R&D Portal - product",
       index: 3,
-      user: this.authorized,
+      user,
     };
 
     const __ = {};
@@ -138,9 +138,15 @@ function* product(next) {
       });
     }
 
+    const product = yield findActiveWorkflowExample(__);
+    const {members} = product;
+
+    const index = members.findIndex(member => member.maill === user.mail);
+
     Object.assign(_, {
       products: __,
-      product: yield findActiveWorkflowExample(__),
+      product,
+      allowedUpload: index === -1 ? false:true,
     });
 
     this.pageName = "platform";